Fixed math error, reduced sky "debris", added options
The variable now called "pc", used in computing x,y when crossing a corner, needs to be pitch/cos(bank), that broke when I moved the /cos out of pitch2pixels().  Now fixed.

I can't find a way to totally remove the left-over bits in the wrong color when tick marks are erased.  New changes minimize that, and also include re-painting the whole "sky" periodically so such "debris" only stays there temporarily.  These problems are due to the white line on the horizon.  In the dark-marks-on-bright-colors color theme that line is not needed, and hence these problems are almost absent.  There is still a minor issue with erasing bank marks that are partly on the horizon line, due to extreme attitude, cannot fully "erase" them by using one color.

Added two options to the horizon screen: dark/bright color theme, and small/large airplane icon.  For now these settings are volatile (not stored in NVS, return to defaults upon booting).  Can make them persistent if and when we know we want to keep them.