no more debris from the horizon line
When the old and new horizon lines cross, "erases" the old by drawing over it twice, in both sky and ground colors, but with clip ranges that limit each to one part of the line.  (Requires setClipRange().)  Also fixed a bug in the bank ticks erase/not-erase logic.  (Lower-bank tick may be skipped due to overlapping with horizon line while higher-bank tick drawn.)  Also improved the overall procedural structure for better efficiency.  Only 800 lines of code to draw the horizon and ticks!